Move freecamera to .app
Also simplify camera and freecamera
Reuse some code between them
