+++ trview.app/trview.app.vcxproj
@@ -23,6 +23,7 @@
     <ClCompile Include="CollapsiblePanel.cpp" />
     <ClCompile Include="Compass.cpp" />
     <ClCompile Include="FileDropper.cpp" />
+    <ClCompile Include="FreeCamera.cpp" />
     <ClCompile Include="ICamera.cpp" />
     <ClCompile Include="ILevelTextureStorage.cpp" />
     <ClCompile Include="Item.cpp" />
@@ -44,6 +45,7 @@
     <ClInclude Include="CollapsiblePanel.h" />
     <ClInclude Include="Compass.h" />
     <ClInclude Include="FileDropper.h" />
+    <ClInclude Include="FreeCamera.h" />
     <ClInclude Include="ICamera.h" />
     <ClInclude Include="ILevelTextureStorage.h" />
     <ClInclude Include="Item.h" />
